# docs/openapi/iam-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Continuum AI IAM API",
    "version": "1.0.0-draft",
    "description": "Review-first contract for authentication, users, projects, project memberships, teams and team memberships. Login, current user, logout and user directory are implemented. Other operations, including refresh, remain proposed. Browser authentication uses an HttpOnly access cookie; bearer access is supported for trusted clients. Access tokens are never returned in browser JSON responses."
  },
...
  "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in",
        "description": "Authenticates an active user, sets continuum_access as an HttpOnly cookie and returns the selected user identity. If the verified account belongs to multiple organizations, returns 409 with eligible organization names and IDs. Invalid credentials use a generic response.",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication succeeded.",
            "headers": {
              "Set-Cookie": {
                "description": "continuum_access; HttpOnly; SameSite=Strict; Path=/; Max-Age=900 (Secure outside local development).",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/OrganizationSelectionRequired" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get the current user",
        "description": "Loads the active user from a valid cookie or bearer access token.",
        "operationId": "getCurrentIdentity",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Authenticated user identity.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CurrentUser" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/iam/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "description": "Returns users visible in the caller's organization scope.",
        "operationId": "listUsers",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/UserStatus" }
          },
          {
            "name": "roleCode",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/RoleCode" }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of users.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/iam/users/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user",
        "operationId": "getUser",
        "parameters": [{ "$ref": "#/components/parameters/UserId" }],
        "responses": {
          "200": {
            "description": "The requested user.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update a user",
        "description": "Updates mutable profile or account-status fields. Only an organization Admin may change status.",
        "operationId": "updateUser",
        "parameters": [{ "$ref": "#/components/parameters/UserId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/iam/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List projects",
        "description": "Returns Project metadata visible to the authenticated User in the trusted Organization. Users with an ACTIVE Organization Membership may read PUBLIC Project metadata; PRIVATE Projects require the documented Project read policy. Organization Admins may list all organization Projects. Nested Project resources are outside this endpoint contract.",
        "operationId": "listProjects",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/ProjectStatus" }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of projects.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a project",
        "description": "Any authenticated User with an ACTIVE Organization Membership in the trusted Organization Context may create a Project in that Organization. New Projects are PRIVATE. An ADMIN or TEAM_LEADER role and a project.create capability grant are not required. Creation bootstraps the creator as an ACTIVE ProjectMembership with a project-scoped MEMBER RoleAssignment, but does not make the creator a Project Leader, owner, or Team. A provisioned current MEMBER Role with project.read is required for the bootstrap; the Project, membership, assignment, and project.create audit event are committed atomically.",
        "operationId": "createProject",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/iam/projects/{projectId}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get a project",
        "description": "Returns Project metadata when the authenticated User has visibility in the trusted Organization. An ACTIVE Organization member may read PUBLIC Project metadata. PRIVATE Projects require the documented Project read policy. Nested Project resources are outside this endpoint contract.",
        "operationId": "getProject",
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "responses": {
          "200": {
            "description": "The requested project.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      },
      "patch": {
        "tags": ["Projects"],
        "summary": "Update a project",
        "operationId": "updateProject",
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "continuum_access",
        "description": "HttpOnly browser access cookie."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access token issued by the login or refresh operation."
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "default": 1, "minimum": 1 }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 100
        }
      },
      "UserId": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/ResourceId" }
      },
      "ProjectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/ResourceId" }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication is missing, invalid, expired or revoked.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated caller lacks permission for the requested scope.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "The addressed resource does not exist or is not visible to the caller.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with an existing unique resource or current state. User status updates can return LAST_ACTIVE_ADMIN or SHARED_ACCOUNT_STATUS_CHANGE.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "OrganizationSelectionRequired": {
        "description": "The password was verified, but an organization must be selected. The details.organizations array contains only organizations assigned to this account.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/OrganizationSelectionError" }
          }
        }
      },
      "ValidationFailed": {
        "description": "The JSON request is understood but violates field or business validation.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "TooManyRequests": {
        "description": "The request exceeded the applicable rate limit.",
        "headers": {
          "Retry-After": {
            "description": "Seconds until the caller may retry.",
            "schema": { "type": "integer", "minimum": 1 }
          },
          "X-RateLimit-Limit": {
            "description": "Maximum requests in the current window.",
            "schema": { "type": "integer", "minimum": 1 }
          },
          "X-RateLimit-Remaining": {
            "description": "Requests remaining in the current window.",
            "schema": { "type": "integer", "minimum": 0 }
          },
          "X-RateLimit-Reset": {
            "description": "UTC Unix timestamp when the current window resets.",
            "schema": { "type": "integer", "minimum": 0 }
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ResourceId": {
        "type": "string",
        "minLength": 1,
        "description": "Opaque resource identifier. Clients must not infer its storage format."
      },
      "UtcTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "Z$",
        "description": "ISO-8601 timestamp normalized to UTC with a trailing Z."
      },
      "RoleCode": {
        "type": "string",
        "enum": ["ADMIN", "TEAM_LEADER", "MEMBER"]
      },
      "UserStatus": {
        "type": "string",
        "enum": ["ACTIVE", "SUSPENDED", "PENDING_INVITE"]
      },
      "ProjectStatus": {
        "type": "string",
        "enum": ["ACTIVE", "ARCHIVED"]
      },
...
      "Pagination": {
        "type": "object",
        "additionalProperties": false,
        "required": ["page", "pageSize", "totalItems", "totalPages"],
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "pageSize": { "type": "integer", "minimum": 1, "maximum": 100 },
          "totalItems": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "details", "requestId"],
        "properties": {
          "code": { "type": "string", "minLength": 1 },
          "message": { "type": "string", "minLength": 1 },
          "details": {
            "type": "object",
            "description": "Machine-readable context. Use an empty object when no additional context is safe to expose.",
            "additionalProperties": true
          },
          "requestId": { "type": "string", "minLength": 1 }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 320 },
          "organizationId": { "$ref": "#/components/schemas/ResourceId" },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024,
            "writeOnly": true
          }
        }
      },
      "CurrentUser": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "email", "name", "organizationId", "roles"],
        "properties": {
          "id": { "$ref": "#/components/schemas/ResourceId" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string", "minLength": 1 },
          "organizationId": { "$ref": "#/components/schemas/ResourceId" },
          "roles": { "type": "array", "items": { "type": "string" }, "uniqueItems": true }
        }
      },
      "LoginResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["user"],
        "properties": {
          "user": { "$ref": "#/components/schemas/CurrentUser" }
        }
      },
...
      "OrganizationSelectionError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "details", "requestId"],
        "properties": {
          "code": { "type": "string", "const": "ORGANIZATION_SELECTION_REQUIRED" },
          "message": { "type": "string" },
          "details": {
            "type": "object",
            "additionalProperties": false,
            "required": ["organizations"],
            "properties": {
              "organizations": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["id", "name"],
                  "properties": {
                    "id": { "$ref": "#/components/schemas/ResourceId" },
                    "name": { "type": "string", "minLength": 1 }
                  }
                }
              }
            }
          },
          "requestId": { "type": "string" }
        }
      },
...
      "User": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "email",
          "fullName",
          "status",
          "roleCodes",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "email": { "type": "string", "format": "email" },
          "fullName": { "type": "string", "minLength": 1, "maxLength": 200 },
          "avatarUrl": { "type": ["string", "null"], "format": "uri" },
          "status": { "$ref": "#/components/schemas/UserStatus" },
          "roleCodes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RoleCode" },
            "uniqueItems": true
          },
          "lastLoginAt": {
            "oneOf": [
              { "$ref": "#/components/schemas/UtcTimestamp" },
              { "type": "null" }
            ]
          },
          "createdAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "updatedAt": { "$ref": "#/components/schemas/UtcTimestamp" }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "fullName": { "type": "string", "minLength": 1, "maxLength": 200 },
          "avatarUrl": { "type": ["string", "null"], "format": "uri" },
          "status": { "$ref": "#/components/schemas/UserStatus" }
        }
      },
      "UserList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data", "pagination"],
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        }
      },
      "Project": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "organizationId",
          "name",
          "code",
          "status",
          "createdBy",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "organizationId": { "$ref": "#/components/schemas/ResourceId" },
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_-]{1,31}$" },
          "description": { "type": ["string", "null"], "maxLength": 2000 },
          "status": { "$ref": "#/components/schemas/ProjectStatus" },
          "createdBy": { "$ref": "#/components/schemas/ResourceId" },
          "createdAt": { "$ref": "#/components/schemas/UtcTimestamp" },
          "updatedAt": { "$ref": "#/components/schemas/UtcTimestamp" }
        }
      },
      "CreateProjectRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "code"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_-]{1,31}$" },
          "description": { "type": "string", "maxLength": 2000 }
        }
      },
      "UpdateProjectRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": ["string", "null"], "maxLength": 2000 },
          "status": { "$ref": "#/components/schemas/ProjectStatus" }
        }
      },
      "ProjectList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data", "pagination"],
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Project" }
          },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        }
      },
...
    }
  }
}